+++ changelog.md
@@ -1,2 +1,1 @@
-- fixed shared conditions not working in all cases on neoforge, added missing conditions there
-- fixed map markers json not working with tags when being sent to clients
+- fixed registration order on fabric, causing issues with components